bsd: merge freebsd and netbsd modules, use sysinfo
