+++ crates/nu-system/Cargo.toml
@@ -19,7 +19,6 @@ workspace = true
 [dependencies]
 nu-utils = { path = "../nu-utils", version = "0.113.0", default-features = false }
 
-libc = { workspace = true }
 log = { workspace = true }
 sysinfo = { workspace = true }
 itertools = { workspace = true }
